requirements.txt
Flask
boto3
flask-cors
serverless-wsgi
passlib
PyJWT